Refactor Chart Symbols into their own list component
